Move subcommand engine classes out of the lobby package, because it's not lobby specific
