+++ engine/src/main/java/games/aternos/odessa/engine/lobby/command/odessa/OdessaCommand.java
@@ -4,7 +4,7 @@ import games.aternos.odessa.engine.lobby.GameLobbySystem;
 import games.aternos.odessa.engine.lobby.command.odessa.subcommand.AddSpawnSubCommand;
 import games.aternos.odessa.engine.lobby.command.odessa.subcommand.CreateArenaSubCommand;
 import games.aternos.odessa.engine.lobby.command.odessa.subcommand.SetLobbySpawnSubCommand;
-import games.aternos.odessa.engine.lobby.command.subcommand.SubCommandDrivenCommand;
+import games.aternos.odessa.engine.subcommand.SubCommandDrivenCommand;
 
 public class OdessaCommand extends SubCommandDrivenCommand {
