+++ engine/src/main/java/games/aternos/odessa/engine/subcommand/SubCommandDrivenCommand.java
@@ -1,4 +1,4 @@
-package games.aternos.odessa.engine.lobby.command.subcommand;
+package games.aternos.odessa.engine.subcommand;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
